requirements/requirements.txt
click~=8.0
click-default-group~=1.2
neon_utils~=1.0
pyyaml~=5.4
